rewrite display loop; add syntax highlighting; add theme
Display loop has been rewritten. It now supports multiple highlighted
sections and multiple selections. It sends as few control sequences to
change color states as possible. The display is still fully redrawn on
every key press, but the speed is reasonable.

A simple go syntax highlighter has been added that parses only the visible
screenful of text (which means multi-line syntax won't be properly
recognized) using the go scanner from the standard go library and prepares
highlighted sections for the display. Even though this happens every
pass (!) through the main loop, which means that the syntax is parsed
every time a key is pressed, it is very fast.

A modest theme has been implemented. It is very basic, but sufficient.